Take care of the ADT unsizing case.
